pyproject.toml
[project]
name = "paper-review-service"
version = "0.2.0"
description = "AI/ML 논문 자동 수집 및 요약 서비스 - arXiv + HuggingFace 통합"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    # Core
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.0",

    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.12",

    # HTTP Client
    "httpx>=0.27.0",
    "requests>=2.32.0",

    # PDF & Image Processing
    "pymupdf>=1.26.0",
    "pillow>=12.0.0",

    # LLM & API
    "arxiv>=2.3.1",

    # Utilities
    "python-dateutil>=2.9.0",
    "tqdm>=4.67.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-pyyaml",
    "types-requests",
]

[project.scripts]
paper-review = "paper_review.cli:main"
paper-review-web = "paper_review.web:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_review"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
